+++ src/focusIsHidden.js
@@ -1,8 +1,11 @@
 import { toArray } from './utils/array';
 import { FOCUS_ALLOW } from './constants';
 
-const focusIsHidden = () => (
+/**
+ * focus is hidden FROM the focus-lock
+ * ie contained inside a node focus-lock shall ignore
+ * @returns {boolean}
+ */
+export const focusIsHidden = () => (
   document && toArray(document.querySelectorAll(`[${FOCUS_ALLOW}]`)).some(node => node.contains(document.activeElement))
 );
-
-export default focusIsHidden;
