+++ src/index.js
@@ -1,10 +1,11 @@
 import tabHook from './tabHook';
-import focusMerge, { getFocusabledIn } from './focusMerge';
+import focusMerge from './focusMerge';
 import focusInside from './focusInside';
-import focusIsHidden from './focusIsHidden';
-import setFocus from './setFocus';
+import { focusIsHidden } from './focusIsHidden';
+import { setFocus } from './setFocus';
 import * as constants from './constants';
 import getAllAffectedNodes from './utils/all-affected';
+import { getFocusabledIn } from './focusables';
 
 export {
   tabHook,
