+++ src/utils/array.js
@@ -1,3 +1,8 @@
+/*
+IE11 support
+ */
+
+
 export const toArray = (a) => {
   const ret = Array(a.length);
   for (let i = 0; i < a.length; ++i) {
